+++ ext/nokogiri/xml_element_content.c
@@ -29,7 +29,7 @@ static VALUE get_type(VALUE self)
   xmlElementContentPtr elem;
   Data_Get_Struct(self, xmlElementContent, elem);
 
-  return INT2NUM(elem->type);
+  return INT2NUM((long)elem->type);
 }
 
 /*
@@ -74,7 +74,7 @@ static VALUE get_occur(VALUE self)
   xmlElementContentPtr elem;
   Data_Get_Struct(self, xmlElementContent, elem);
 
-  return INT2NUM(elem->ocur);
+  return INT2NUM((long)elem->ocur);
 }
 
 /*
